+++ designer/client/src/components/graph/node-modal/ScenarioProperty.tsx
@@ -5,7 +5,7 @@ import { ExpressionLang } from "./editors/expression/types";
 import { errorValidator, PossibleValue } from "./editors/Validators";
 import { NodeType, NodeValidationError } from "../../../types";
 
-export interface AdditionalPropertyConfig {
+export interface ScenarioPropertyConfig {
     editor: any;
     label: string;
     values: Array<PossibleValue>;
@@ -15,7 +15,7 @@ interface Props {
     showSwitch: boolean;
     showValidation: boolean;
     propertyName: string;
-    propertyConfig: AdditionalPropertyConfig;
+    propertyConfig: ScenarioPropertyConfig;
     propertyErrors: NodeValidationError[];
     editedNode: NodeType;
     onChange: <K extends keyof NodeType>(property: K, newValue: NodeType[K], defaultValue?: NodeType[K]) => void;
@@ -23,7 +23,7 @@ interface Props {
     readOnly: boolean;
 }
 
-export default function AdditionalProperty(props: Props) {
+export default function ScenarioProperty(props: Props) {
     const { showSwitch, showValidation, propertyName, propertyConfig, propertyErrors, editedNode, onChange, renderFieldLabel, readOnly } =
         props;
 
@@ -48,7 +48,7 @@ export default function AdditionalProperty(props: Props) {
             key={propertyName}
             showSwitch={showSwitch}
             showValidation={showValidation}
-            //AdditionalProperties do not use any variables
+            //ScenarioProperties do not use any variables
             variableTypes={{}}
             validators={validators}
         />