+++ designer/server/src/main/scala/pl/touk/nussknacker/engine/ProcessingTypeData.scala
@@ -3,7 +3,7 @@ package pl.touk.nussknacker.engine
 import _root_.sttp.client3.SttpBackend
 import akka.actor.ActorSystem
 import com.typesafe.config.Config
-import pl.touk.nussknacker.engine.api.component.AdditionalPropertyConfig
+import pl.touk.nussknacker.engine.api.component.ScenarioPropertyConfig
 import pl.touk.nussknacker.engine.api.deployment.{DeploymentManager, ProcessingTypeDeploymentService}
 import pl.touk.nussknacker.engine.definition.DefinitionExtractor.ObjectDefinition
 import pl.touk.nussknacker.engine.definition.ProcessDefinitionExtractor.ProcessDefinition
@@ -17,7 +17,7 @@ final case class ProcessingTypeData private (
     modelData: ModelData,
     staticObjectsDefinition: ProcessDefinition[ObjectDefinition],
     metaDataInitializer: MetaDataInitializer,
-    additionalPropertiesConfig: Map[String, AdditionalPropertyConfig],
+    scenarioPropertiesConfig: Map[String, ScenarioPropertyConfig],
     additionalValidators: List[CustomProcessValidator],
     usageStatistics: ProcessingTypeUsageStatistics
 ) {
@@ -66,9 +66,9 @@ object ProcessingTypeData {
   ): ProcessingTypeData = {
     import net.ceedubs.ficus.Ficus._
     import pl.touk.nussknacker.engine.util.config.FicusReaders._
-    val additionalProperties =
-      deploymentManagerProvider.additionalPropertiesConfig(managerConfig) ++ modelData.processConfig
-        .getOrElse[Map[String, AdditionalPropertyConfig]]("additionalPropertiesConfig", Map.empty)
+    val scenarioProperties =
+      deploymentManagerProvider.scenarioPropertiesConfig(managerConfig) ++ modelData.processConfig
+        .getOrElse[Map[String, ScenarioPropertyConfig]]("scenarioPropertiesConfig", Map.empty)
 
     val metaDataInitializer = deploymentManagerProvider.metaDataInitializer(managerConfig)
     val staticObjectsDefinition =
@@ -79,7 +79,7 @@ object ProcessingTypeData {
       modelData,
       staticObjectsDefinition,
       metaDataInitializer,
-      additionalProperties,
+      scenarioProperties,
       deploymentManagerProvider.additionalValidators(managerConfig),
       ProcessingTypeUsageStatistics(managerConfig)
     )