+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/DefinitionResources.scala
@@ -54,7 +54,7 @@ class DefinitionResources(
                     fragments,
                     isFragment,
                     processCategoryService,
-                    processingTypeData.additionalPropertiesConfig,
+                    processingTypeData.scenarioPropertiesConfig,
                     processingType,
                     additionalComponentsUIConfigProvider
                   )