+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/validation/ScenarioPropertiesValidator.scala
@@ -5,36 +5,36 @@ import cats.data.Validated.{Invalid, Valid, invalid, valid}
 import pl.touk.nussknacker.engine.api.context.PartSubGraphCompilationError
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.{MissingRequiredProperty, UnknownProperty}
 import pl.touk.nussknacker.engine.api.definition.{MandatoryParameterValidator, ParameterValidator}
-import pl.touk.nussknacker.engine.api.component.AdditionalPropertyConfig
+import pl.touk.nussknacker.engine.api.component.ScenarioPropertyConfig
 import pl.touk.nussknacker.engine.api.NodeId
 import pl.touk.nussknacker.restmodel.displayedgraph.DisplayableProcess
 import pl.touk.nussknacker.restmodel.validation.PrettyValidationErrors
 import pl.touk.nussknacker.restmodel.validation.ValidationResults.ValidationResult
-import pl.touk.nussknacker.ui.definition.additionalproperty.AdditionalPropertyValidatorDeterminerChain
+import pl.touk.nussknacker.ui.definition.scenarioproperty.ScenarioPropertyValidatorDeterminerChain
 import pl.touk.nussknacker.ui.process.processingtypedata.ProcessingTypeDataProvider
 
-class AdditionalPropertiesValidator(
-    additionalPropertiesConfig: ProcessingTypeDataProvider[Map[String, AdditionalPropertyConfig], _]
+class ScenarioPropertiesValidator(
+    scenarioPropertiesConfig: ProcessingTypeDataProvider[Map[String, ScenarioPropertyConfig], _]
 ) {
 
   import cats.implicits._
 
   implicit val nodeId: NodeId = NodeId("properties")
 
-  type PropertyConfig = Map[String, AdditionalPropertyConfig]
+  type PropertyConfig = Map[String, ScenarioPropertyConfig]
 
   def validate(process: DisplayableProcess): ValidationResult =
-    additionalPropertiesConfig.forType(process.processingType) match {
+    scenarioPropertiesConfig.forType(process.processingType) match {
       case None =>
         ValidationResult.globalErrors(List(PrettyValidationErrors.noValidatorKnown(process.processingType)))
 
       case Some(config) => {
-        val additionalProperties = process.properties.additionalFields.properties.toList
+        val scenarioProperties = process.properties.additionalFields.properties.toList
 
         val validated = (
-          getConfiguredValidationsResults(config, additionalProperties),
-          getMissingRequiredPropertyValidationResults(config, additionalProperties),
-          getUnknownPropertyValidationResults(config, additionalProperties)
+          getConfiguredValidationsResults(config, scenarioProperties),
+          getMissingRequiredPropertyValidationResults(config, scenarioProperties),
+          getUnknownPropertyValidationResults(config, scenarioProperties)
         )
           .mapN { (_, _, _) => () }
 
@@ -47,14 +47,14 @@ class AdditionalPropertiesValidator(
       }
     }
 
-  private def getConfiguredValidationsResults(config: PropertyConfig, additionalProperties: List[(String, String)]) = {
+  private def getConfiguredValidationsResults(config: PropertyConfig, scenarioProperties: List[(String, String)]) = {
     val validatorsByPropertyName = config
       .map(propertyConfig =>
-        propertyConfig._1 -> AdditionalPropertyValidatorDeterminerChain(propertyConfig._2).determine()
+        propertyConfig._1 -> ScenarioPropertyValidatorDeterminerChain(propertyConfig._2).determine()
       )
 
     val propertiesWithConfiguredValidator = for {
-      property  <- additionalProperties
+      property  <- scenarioProperties
       validator <- validatorsByPropertyName.getOrElse(property._1, List.empty)
     } yield (property, config.get(property._1), validator)
 
@@ -68,13 +68,13 @@ class AdditionalPropertiesValidator(
 
   private def getMissingRequiredPropertyValidationResults(
       config: PropertyConfig,
-      additionalProperties: List[(String, String)]
+      scenarioProperties: List[(String, String)]
   ) = {
     config
       .filter(_._2.validators.nonEmpty)
       .filter(_._2.validators.get.contains(MandatoryParameterValidator))
       .map(propertyConfig =>
-        (propertyConfig._1, propertyConfig._2, MissingRequiredPropertyValidator(additionalProperties.map(_._1)))
+        (propertyConfig._1, propertyConfig._2, MissingRequiredPropertyValidator(scenarioProperties.map(_._1)))
       )
       .toList
       .map { case (propertyName, config, validator) =>
@@ -86,9 +86,9 @@ class AdditionalPropertiesValidator(
 
   private def getUnknownPropertyValidationResults(
       config: PropertyConfig,
-      additionalProperties: List[(String, String)]
+      scenarioProperties: List[(String, String)]
   ) = {
-    additionalProperties
+    scenarioProperties
       .map(property => (property._1, UnknownPropertyValidator(config)))
       .map { case (propertyName, validator) =>
         validator.isValid(propertyName).toValidatedNel
@@ -110,7 +110,7 @@ private final case class MissingRequiredPropertyValidator(actualPropertyNames: L
 
 }
 
-private final case class UnknownPropertyValidator(config: Map[String, AdditionalPropertyConfig]) {
+private final case class UnknownPropertyValidator(config: Map[String, ScenarioPropertyConfig]) {
 
   def isValid(propertyName: String)(implicit nodeId: NodeId): Validated[PartSubGraphCompilationError, Unit] = {