+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 1.13.0 (Not released yet)
 -------------------------
+* [#4860](https://github.com/TouK/nussknacker/pull/4860) Rename `additionalProperties` to `scenarioProperties`
 * [#4711](https://github.com/TouK/nussknacker/pull/4711) Added AdditionalComponentsUIConfigProviderFactory API that allows changing components' configs without model reload
 
 1.12.3 (26 Oct 2023)